Inline daily_report helpers into pm_audio_briefing.daily_log; delete daily_report.py
Closes deferred item #1 from issue #1292.

The legacy reflections/daily_report.py module owned the aggregator,
renderer, and vault writer that powered the daily-report-and-notify
registry entry. PR #1281 made the new pm-audio-briefing slot
dispatcher wrap those helpers, leaving the original module as a transitive
import chain. This commit inlines all the helpers into the slot module
(pm_audio_briefing/daily_log.py) and deletes daily_report.py.

The legacy run() function and its delivery helpers
(_select_target_chat_id, _enqueue_voice_note, _send_audio_brief,
_build_audio_brief) were *not* inlined — they were dead once the slot
dispatcher took over delivery. Their tests were correspondingly retired:
- tests/integration/reflections/test_daily_report_integration.py — DELETED
  (legacy run() end-to-end; behavior covered by test_pm_briefings_dispatch.py)
- tests/unit/reflections/test_daily_report_*.py — RENAMED to test_daily_log_*
  with imports updated; chat-selection tests removed (delivery now lives in
  the dispatcher's delivery.py)

The legacy registry entry daily-report-and-notify (callable
reflections.daily_report.run) must be flipped to enabled: false in
the vault yaml as part of the cutover (deferred item #4); see PR body for
the operator step.
